update aoe catalog compaction
